Revert bridge epochification
Revert r363492, r363491, r363430, r363429 and r362650.

The introduction of epoch in the network stack is incomplete in stable/12, and
there are simply too many limitations to make the bridge epoch code work there.

The final problem is capability configuration of the bridge member interfaces.
if_bridge needs to enable promiscuous mode, which for certain drivers (e1000
for example) can sleep. In stable/12 we may not sleep within epoch.


git-svn-id: https://svn.freebsd.org/base/stable/12@363568 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f