MFC r363383: dtrace/fbt: fix return probe arguments on arm
arg0 should be an offset of the return point within the function, arg1
should be the return value.  Previously the return probe had arguments as
if for the entry probe.

Tested on armv7.

andrew noted that the same problem seems to be present on arm64, mips,
and riscv.
I am not sure if I will get around to fixing those.  So, platform users
or anyone looking to make a contribution please be aware of this
opportunity.


git-svn-id: https://svn.freebsd.org/base/stable/12@363697 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f