Import version 3.1.4
This version makes dc exit after processing all commands passed via -e or -f
instead of waiting for more input on STDIN (add "-f -" to the command line
to emulate the behavior of versionm 3.1.3 and earlier, if desired).

The version and copyright message are no longer printed for interactive
sessions as was the case with the prior implementation in the FreeBSD base
system.

Obtained from:	https://git.yzena.com/gavin/bc


git-svn-id: https://svn.freebsd.org/base/vendor/bc/dist@363807 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f