Upgrade	to version 3.1.4
This version omits the printing	of a copyright header in interactive mode
and the dc command now exits after	execution of	the commands passed via	-e
or -f instead of switching to interactive mode.	To pass	further	commands
via STDIN when dc has been invoked with	-e or -f, add "-f -" to	the
parameter list.


git-svn-id: https://svn.freebsd.org/base/head@363810 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f