Connect the tests provided with the new bc and dc
The tests compare the command output (including of error cases) with the
expected output and exit code.

Not all tests are executed, since some expect to have a known good bc and
dc binary installed and compare results of large amounts of generated data
being processed by both versions to test for regressions.


git-svn-id: https://svn.freebsd.org/base/head@363811 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f