MFC r361257: bsdinstall: do a `certctl rehash` upon installation of config
If certctl is installed on the system we're configuring, do a certctl
rehash.

Note that certctl may not be present if the world we've installed was built
either WITHOUT_OPENSSL or WITHOUT_CAROOT. In this scenario, we don't
currently see if the host has a certctl as this may be an indication that
the system *shouldn't* have certs installed into /etc/ssl.


git-svn-id: https://svn.freebsd.org/base/stable/12@363913 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f