pkgbase: Add a new variable PKG_TIMESTAMP
libarchive uses the SOURCE_DATE_EPOCH env variable to set the date of file
in an archive, this is useful for reproducibility.
Add a variable name PKG_TIMESTAMP that take a epoch time and set SOURCE_DATE_EPOCH
to this. By default it is the current time so no changes here.


git-svn-id: https://svn.freebsd.org/base/head@363964 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f