crunchgen: use pwd -P without env
The -P flag is required by POSIX so we don't have to care whether pwd is
a shell builtin or not. This also allows removing pwd from the list of
bootstrap tools since all shells we care about for building have a
builtin pwd command. This effectively reverts r364190.

Suggested By:	rgrimes, jrtc27


git-svn-id: https://svn.freebsd.org/base/head@364234 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f