MFC r364028: Implement radix_tree_store() in the LinuxKPI for use with the coming extensible arrays implementation.
While at it add some more comments explaining the current
radix_tree_insert() function and make sure to clean the root node when
the radix tree reaches the maximum height. This can happen if the
index passed is too big when the tree is empty.

The radix_tree_store() function is basically a copy of the
radix_tree_insert() function with some added functionality.

The radix_tree_store() function is local to FreeBSD and does not yet
exist in Linux.

Reviewed by:		kib
Sponsored by:		Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/11@364383 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f